Add some more informative text on main page.
This is text adapted from a article appearing in June 2008 ATC News for
Meridian Energy. An older and more similar OCRed, markdowned and edited
version is in ats.html.

Adedd a couple of not-very good photos. I wanted to put them into seperate
left-float blocks so the don't break up the text so much on narrower screens
but on my test browser the lack of text meant that only one long line ended up
between the two images. Ugly. Anyway, the captions break up the two images a
bit.