Initial tentative conversion to use jekyll for prose.io.
I cloned this repo <https://github.com/prose/starter> and modified it a bit.
Mostly just to take away some complexity. There is still a bunch of stuff that
can be cut but I have spent enough time screwing around with it locally and
want to see if it works.

Took me like an hour to figure out that the reason the individual content
files weren't being included is because they have to be named magiaclly, yay
ruby!